adds functionality to tank plugin - drivable tank robot
