+++ 11_Robots/ghost_high_stakes/config/robot_hardware_config_tank.yaml
@@ -34,7 +34,7 @@ port_configuration:
             port: 10
             type: MOTOR
             reversed: true
-            config: drive_motor_config
+            config: indexer_config
         drive_rttf:
             port: 16
             type: MOTOR
