+++ 11_Robots/ghost_high_stakes/package.xml
@@ -13,7 +13,7 @@
   <test_depend>ament_lint_common</test_depend>
 
   <exec_depend>ghost_ros_interfaces</exec_depend>
-  <exec_depend>ghost_swerve</exec_depend>
+  <exec_depend>ghost_tank</exec_depend>
   <exec_depend>ghost_localization</exec_depend>
   <exec_depend>ghost_sensing</exec_depend>
   <exec_depend>realsense_camera</exec_depend>
