+++ 11_Robots/ghost_tank/include/ghost_tank/bt_nodes/moveToPoseTrapezoid.hpp
@@ -19,7 +19,7 @@ public:
 	// If your Node has ports, you must use this constructor signature
 	MoveToPoseTrapezoid(const std::string& name, const BT::NodeConfig& config, std::shared_ptr<rclcpp::Node> node_ptr,
 	           std::shared_ptr<ghost_v5_interfaces::RobotHardwareInterface> rhi_ptr,
-	           std::shared_ptr<tankModel> tank_ptr);
+	           std::shared_ptr<TankModel> tank_ptr);
 
 	// It is mandatory to define this STATIC method.
 	static BT::PortsList providedPorts();
@@ -43,7 +43,7 @@ private:
 	T get_input(std::string key);
 
 	std::shared_ptr<ghost_v5_interfaces::RobotHardwareInterface> rhi_ptr_;
-	std::shared_ptr<tankModel> tank_ptr_;
+	std::shared_ptr<TankModel> tank_ptr_;
 	rclcpp::Publisher<ghost_msgs::msg::DrivetrainCommand>::SharedPtr command_pub_;
 	std::chrono::time_point<std::chrono::system_clock> start_time_;
 	std::chrono::time_point<std::chrono::system_clock> plan_time_;
