+++ 11_Robots/ghost_tank/include/ghost_tank/cubic_motion_planner.hpp
