+++ 11_Robots/ghost_tank/include/ghost_tank/tank_tree.hpp
@@ -22,46 +22,42 @@
  */
 
 #pragma once
-
-#include "behaviortree_cpp/behavior_tree.h"
+#include "behaviortree_cpp/bt_factory.h"
+#include "bt_nodes/checkForRestart.hpp"
+#include "bt_nodes/intakeCmd.hpp"
+#include "bt_nodes/loggingNode.hpp"
+#include "bt_nodes/moveToPoseCubic.hpp"
+#include "bt_nodes/autoDone.hpp"
+#include "bt_nodes/autonTimer.hpp"
 #include "ghost_tank/tank_model.hpp"
 #include "ghost_v5_interfaces/robot_hardware_interface.hpp"
 #include "rclcpp/rclcpp.hpp"
+#include "std_msgs/msg/bool.hpp"
+
+// file that contains the custom nodes definitions
+// #include "dummy_nodes.h"
+// using namespace DummyNodes;
 
 namespace ghost_tank
 {
 
-class SwipeTail : public BT::StatefulActionNode
+class TankTree
 {
 public:
-  SwipeTail(
-    const std::string & name, const BT::NodeConfig & config, std::shared_ptr<rclcpp::Node> node_ptr,
-    std::shared_ptr<ghost_v5_interfaces::RobotHardwareInterface> rhi_ptr,
-    std::shared_ptr<tankModel> tank_ptr);
-
-  // It is mandatory to define this STATIC method.
-  static BT::PortsList providedPorts();
-
-  /// Method called once, when transitioning from the state IDLE.
-  /// If it returns RUNNING, this becomes an asynchronous node.
-  BT::NodeStatus onStart();
-
-  /// method invoked when the action is already in the RUNNING state.
-  BT::NodeStatus onRunning();
-
-  /// when the method halt() is called and the action is RUNNING, this method is invoked.
-  /// This is a convenient place todo a cleanup, if needed.
-  void onHalted();
+	TankTree(std::string bt_path,
+			   std::string bt_path_interaction,
+	           std::shared_ptr<ghost_v5_interfaces::RobotHardwareInterface> robot_hardware_interface_ptr,
+	           std::shared_ptr<TankModel> tank_ptr,
+	           std::shared_ptr<rclcpp::Node> node_ptr);
+	void tick_tree();
+	void tick_tree_interaction();
 
 private:
-  template<typename T>
-  T get_input(std::string key);
-
-  std::shared_ptr<ghost_v5_interfaces::RobotHardwareInterface> rhi_ptr_;
-  std::shared_ptr<tankModel> tank_ptr_;
-  bool started_;
+  std::string bt_path_;
+  std::string bt_path_interaction_;
+  BT::Tree tree_;
+  BT::Tree tree_interaction_;
   std::shared_ptr<rclcpp::Node> node_ptr_;
-  std::chrono::time_point<std::chrono::system_clock> start_time_;
 };
 
 } // namespace ghost_tank
