+++ 11_Robots/ghost_tank/include/ghost_tank/trapezoid_motion_planner.hpp
