+++ 11_Robots/ghost_tank/src/bt_nodes/intakeCmd.cpp
@@ -28,20 +28,20 @@ namespace ghost_tank
 
 IntakeCmd::IntakeCmd(const std::string& name, const BT::NodeConfig& config, std::shared_ptr<rclcpp::Node> node_ptr,
                      std::shared_ptr<ghost_v5_interfaces::RobotHardwareInterface> rhi_ptr,
-                     std::shared_ptr<tankModel> tank_ptr) :
+                     std::shared_ptr<TankModel> tank_ptr) :
 	BT::SyncActionNode(name, config),
 	node_ptr_(node_ptr),
 	rhi_ptr_(rhi_ptr),
 	tank_ptr_(tank_ptr){
 	// ros params
 	// node_ptr_->declare_parameter("tank_robot_plugin.burnout_absolute_velocity_threshold_rpm", 50.0);
-	burnout_absolute_rpm_threshold_        = node_ptr_->get_parameter("tank_robot_plugin.burnout_absolute_velocity_threshold_rpm").as_double();
+	// burnout_absolute_rpm_threshold_        = node_ptr_->get_parameter("tank_robot_plugin.burnout_absolute_velocity_threshold_rpm").as_double();
 
 	// node_ptr_->declare_parameter("tank_robot_plugin.burnout_stall_duration_ms", 1000);
-	burnout_stall_duration_ms_    = node_ptr_->get_parameter("tank_robot_plugin.burnout_stall_duration_ms").as_int();
+	// burnout_stall_duration_ms_    = node_ptr_->get_parameter("tank_robot_plugin.burnout_stall_duration_ms").as_int();
 
 	// node_ptr_->declare_parameter("tank_robot_plugin.burnout_cooldown_duration_ms", 1000);
-	burnout_cooldown_duration_ms_ = node_ptr_->get_parameter("tank_robot_plugin.burnout_cooldown_duration_ms").as_int();
+	// burnout_cooldown_duration_ms_ = node_ptr_->get_parameter("tank_robot_plugin.burnout_cooldown_duration_ms").as_int();
 }
 
 // It is mandatory to define this STATIC method.
